refactor(web):  fsd 구조에 맞게 모임 상세 페이지 관련 폴더 구조 정리
