+++ apps/web/src/_pages/moim-detail/ui/information-container.tsx
@@ -4,11 +4,11 @@ import { MoreHorizontal } from "@moum-zip/ui/icons";
 import { Button, Dropdown, Tag } from "@ui/components";
 import { cn } from "@ui/lib/utils";
 import { useState } from "react";
-import CrownIcon from "../assets/svg/crown.svg";
-import LocationIcon from "../assets/svg/location.svg";
-import type { InformationData } from "../model/types";
-import { AlertModal } from "./alert-modal";
-import { LikeButton } from "./like-button";
+import CrownIcon from "@/_pages/moim-detail/assets/crown.svg";
+import LocationIcon from "@/_pages/moim-detail/assets/location.svg";
+import { AlertModal } from "@/_pages/moim-detail/ui/alert-modal";
+import type { InformationData } from "@/entities/moim-detail/model/types";
+import { LikeButton } from "@/features/moim-detail/ui/like-button";
 
 type ViewType = "member" | "manager";
