+++ apps/web/src/_pages/moim-detail/ui/moim-detail-client.tsx
@@ -4,8 +4,8 @@ import { useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
 import { CompactCard, DescriptionSection, InformationContainer, PersonnelContainer } from "@/_pages/moim-detail";
 import { deleteMeetingAction, favoriteMeetingAction, joinMeetingAction } from "@/_pages/moim-detail/actions";
-import LocationIcon from "@/_pages/moim-detail/assets/svg/location.svg";
-import type { InformationData, PersonnelData, RecommendedMeetingData } from "@/_pages/moim-detail/model/types";
+import LocationIcon from "@/_pages/moim-detail/assets/location.svg";
+import type { InformationData, PersonnelData, RecommendedMeetingData } from "@/entities/moim-detail/model/types";
 import { ROUTES } from "@/shared/config/routes";
 
 interface MoimDetailClientProps {
