+++ apps/web/src/features/moim-detail/use-cases/delete-meeting.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it, vi } from "vitest";
-import { deleteMeeting } from "./delete-meeting";
+import { deleteMeeting } from "@/features/moim-detail/use-cases/delete-meeting";
 
 describe("deleteMeeting", () => {
   it("meetingId로 모임 삭제를 요청한다", async () => {
